refactor: move session handlers to session file
